test(color-selector): move interaction tests to browser mode
